requirements.txt
asgiref==3.3.1
Django==3.1.4
django-mathfilters==1.0.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.15
Pillow==8.0.1
pytz==2020.5
sqlparse==0.4.1
gunicorn==20.0.4
django-modeltranslation==0.16.1
pytils~=0.3
djangorestframework~=3.12.2
django-cors-headers~=3.7.0